+++ madsim/src/sim/mod.rs
@@ -7,6 +7,7 @@ pub(crate) use self::runtime::context;
 #[cfg_attr(docsrs, doc(cfg(feature = "macros")))]
 pub use madsim_macros::{main, test, tokio_main, tokio_test};
 
+pub mod buggify;
 mod config;
 pub mod fs;
 pub mod net;
@@ -15,6 +16,7 @@ pub mod plugin;
 pub mod rand;
 #[cfg_attr(docsrs, doc(cfg(madsim)))]
 pub mod runtime;
+pub mod signal;
 pub mod task;
 pub mod time;
 mod utils;