+++ madsim/src/std/signal.rs
@@ -1,0 +1,3 @@
+//! Asynchronous signal handling.
+
+pub use tokio::signal::ctrl_c;