+++ CHANGELOG.md
@@ -6,6 +6,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- madsim: Add `buggify`.
+- madsim: Add `JoinHandle::{id, is_finished}`.
+- madsim: Add `signal::ctrl_c` and `Handle::send_ctrl_c`.
+- madsim: Add `Handle::is_exit`.
+
+### Changed
+
+- madsim: After the initial task completes, the other tasks of the node are dropped.
+
+### Fixed
+
+- etcd: Return error on "lease not found".
+
 ## [0.2.14] - 2023-01-30
 
 ### Added